+++ pages/index.tsx
@@ -1,13 +1,21 @@
 import { AddUrlBox } from "../src/components/AddUrlBox";
 import { Layout } from "../src/components/Layout";
 import { UrlCards } from "../src/components/UrlCards";
-import { URLCARDDATA2, URLCARDDATA1 } from "../src/data/mockData";
+import { db } from "../src/firebase/firebase.admin";
+import { collection } from "firebase/firestore";
 // add it to firebase & apis
 import type { NextPage } from "next";
 import Head from "next/head";
 import Image from "next/image";
+import { useCollectionData } from "react-firebase-hooks/firestore";
 
 const Home: NextPage = () => {
+  const URLCARDDATA1: string[] = [];
+  const colRef = collection(db, "links");
+  const [docs, loading, error] = useCollectionData(colRef);
+  // categories
+  console.log("fetched data!!", docs[0]?.name);
+
   return (
     <Layout>
       {/* SEARCH BOX */}
@@ -17,8 +25,9 @@ const Home: NextPage = () => {
       {/* ADD BUTTON */}
       <AddUrlBox />
       {/* URL CARDS */}
-      <UrlCards cardsData={URLCARDDATA1} />
-      <UrlCards cardsData={URLCARDDATA2} />
+      <div className="bg-red-200 w-full h-screen">
+        <UrlCards cardsData={URLCARDDATA1} />
+      </div>
     </Layout>
   );
 };